% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_census}
\alias{us_census}
\title{Records for individuals on titanic.}
\format{
A data frame with 2201 rows and 5 variables:
\describe{
\item{id}{an arbitrary id for each individual}
\item{class}{1st, 2nd, 3rd for each passenger}
\item{sex}{Male, Female for each passenger}
\item{age}{Child, Adult for each passenger}
\item{survived}{No, Yes for each passenger}
...
}
}
\source{
A version of the Titanic data from the datasets package
}
\usage{
us_census
}
\description{
A dataset with a row for each individual on the titanic
}
\keyword{datasets}
