% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R, R/geom_sf_county.R
\name{geom_sf_county}
\alias{geom_sf_county}
\title{Title}
\usage{
geom_sf_county(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

geom_sf_county(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title

Title
}
\examples{
library(ggplot2)
us_census |>
ggplot() +
  coord_sf() +
  aes(fips = FIPS) +
  geom_sf_county(state = "Illinois") +
  geom_sf_county(state = "Illinois",
                 county = c("Cook", "Champaign"),
                 color = "red", linewidth = 1) +
  aes(fill = mean_work_travel) +
  scale_fill_viridis_c(option = "magma")

last_plot() +
  aes(fill = per_capita_income) +
  scale_fill_viridis_c()
library(ggplot2)
us_census |>
ggplot() +
  coord_sf() +
  aes(fips = fips) +
  geom_sf_county(state = "Illinois") +
  geom_sf_county(state = "Illinois",
                 county = c("Cook", "Champaign"),
                 color = "red", linewidth = 1) +
  aes(fill = mean_work_travel) +
  scale_fill_viridis_c(option = "magma")

last_plot() +
  aes(fill = per_capita_income) +
  scale_fill_viridis_c()
}
